Add comment edit and delete endpoints to EventController and OfferingController
