+++ eventPlanner/src/main/java/com/ftn/iss/eventPlanner/controller/ProductController.java
@@ -265,20 +265,4 @@ public class ProductController {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @PutMapping(value = "/{id}/comments/{id2}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UpdatedCommentDTO> updateComment(@RequestBody UpdateCommentDTO comment, @PathVariable int id, @PathVariable int id2)
-            throws Exception {
-        UpdatedCommentDTO updatedComment = new UpdatedCommentDTO();
-
-        updatedComment.setId(id2);
-        updatedComment.setContent(comment.getContent());
-        updatedComment.setStatus(comment.getStatus());
-
-        return new ResponseEntity<UpdatedCommentDTO>(updatedComment, HttpStatus.OK);
-    }
-
-    @DeleteMapping(value = "/{id}/comments/{id2}")
-    public ResponseEntity<?> deleteComment(@PathVariable("id") int id, @PathVariable int id2) throws Exception {
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
 }
